Initial code for supporting Stiff ODEs
Start of an implicit midpoint method, which is "simple" and has
order h^2 error. It requires that we implement a root-finding method as
all implicit methods do. We use Newton's Method for simplicity right
now. That in turn requires that we implement numerical derivatives.

The current goal is for the internals of Math::ODE to try the stiff
integration routine if under/overflow is detected.